refactor(sandbox-fc): merge trait adapters into their parent modules
- Merge snapshot_provider.rs into snapshot.rs (FirecrackerSnapshotProvider)
- Merge sandbox_control.rs into control.rs (FirecrackerControl)

Both adapter files were thin wrappers (~50-145 lines) that only
implemented a sandbox trait by delegating to the parent module's
functions. Colocating them reduces file count and makes the
relationship explicit.
