+++ crates/sandbox-fc/src/lib.rs
@@ -11,12 +11,11 @@ mod prerequisites;
 mod process;
 mod runtime;
 mod sandbox;
-mod sandbox_control;
 mod snapshot;
-mod snapshot_provider;
 
 pub use api::{ApiClient, ApiError, BalloonStatistics};
 pub use config::{FirecrackerConfig, SnapshotConfig};
+pub use control::FirecrackerControl;
 pub use factory::{FirecrackerFactory, PREWARM_SCRIPT, config_hash};
 pub use network::{NetnsPool, NetnsPoolConfig};
 pub use paths::{
@@ -24,6 +23,4 @@ pub use paths::{
 };
 pub use runtime::{FirecrackerRuntime, FirecrackerRuntimeProvider};
 pub use sandbox::FirecrackerSandbox;
-pub use sandbox_control::FirecrackerControl;
-pub use snapshot::{SnapshotError, create_snapshot};
-pub use snapshot_provider::FirecrackerSnapshotProvider;
+pub use snapshot::{FirecrackerSnapshotProvider, SnapshotError, create_snapshot};
